Modified `brew autoremove` to remove recursive unused dependencies.
So far, it only checks a package's required dependencies to find out which ones to remove. Adding support for build and recommended should be trivial. For this patch to work, you need to `edit` your formulas so that only immediate required dependencies are listed as required dependencies. Ex - `wget` has `libunistring` listed as a direct dependency even though it isn't.

To use, pass the package you want to remove as an argument to `brew autoremove`. Ex. `brew autoremove wget`.